+++ TODO.md
@@ -1,7 +1,6 @@
 # TODO
 
 - dark mode
-- sorting for setlists, songs, and feels
 - better remove member from band settings (should know if you are removing self)
 - fix links in song form
 - drag/drop whole sets
